Finish removing Triton from workflow runners
